requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
apscheduler==3.9.1
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
mysqlclient==2.1.0
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via apscheduler
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
six==1.16.0
    # via apscheduler
sqlalchemy==1.4.37
    # via -r requirements.in
tzdata==2022.1
    # via tzlocal
tzlocal==4.2
    # via apscheduler

# The following packages are considered to be unsafe in a requirements file:
# setuptools
